# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WeatherData {
  city: string;
  country: string;
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  description: string;
  icon: string;
  visibility: number;
  pressure: number;
}

export type CurrencyRatesRates = { [key: string]: number };

export interface CurrencyRates {
  base: string;
  rates: CurrencyRatesRates;
  timestamp: number;
}

export interface ShortenUrlRequest {
  url: string;
}

export interface ShortenUrlResponse {
  originalUrl: string;
  shortUrl: string;
  shortCode: string;
}

export interface TextRequest {
  text: string;
}

export interface TextResponse {
  original: string;
  result: string;
}

export type AiDetectionResultSentencesItem = {
  text: string;
  score: number;
};

export interface AiDetectionResult {
  aiScore: number;
  humanScore: number;
  verdict: string;
  sentences: AiDetectionResultSentencesItem[];
}

export type PlagiarismResultMatchesItem = {
  text: string;
  similarity: number;
};

export interface PlagiarismResult {
  score: number;
  verdict: string;
  uniqueScore: number;
  matches: PlagiarismResultMatchesItem[];
}

export interface ImageBase64Request {
  /** Base64 encoded image */
  imageBase64: string;
  /** Image MIME type (image/jpeg, image/png) */
  mimeType: string;
}

export interface EraseObjectRequest {
  /** Base64 encoded original image */
  imageBase64: string;
  /** Base64 encoded mask image (white = erase, black = keep) */
  maskBase64: string;
  mimeType: string;
}

export interface ImageBase64Response {
  /** Base64 encoded result image */
  resultBase64: string;
  mimeType: string;
}

export interface BlogPostSummary {
  slug: string;
  title: string;
  excerpt: string;
  category: string;
  author: string;
  publishedAt: string;
  readTime: number;
  coverImage: string;
  tags: string[];
}

export interface BlogPost {
  slug: string;
  title: string;
  excerpt: string;
  /** HTML content of the post */
  content: string;
  category: string;
  author: string;
  publishedAt: string;
  readTime: number;
  coverImage: string;
  tags: string[];
}

export interface BlogPostList {
  posts: BlogPostSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export type GetWeatherParams = {
  city: string;
};

export type GetWeatherByCoordsParams = {
  lat: number;
  lon: number;
};

export type GetCurrencyRatesParams = {
  base: string;
};

export type GetBlogPostsParams = {
  page?: number;
  limit?: number;
  category?: string;
};