# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  slug: string;
  title: string;
  excerpt: string;
  /** HTML content of the post */
  content: string;
  category: string;
  author: string;
  publishedAt: string;
  readTime: number;
  coverImage: string;
  tags: string[];
}