# lib/api-zod/src/generated/types/blogPostList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostSummary } from "./blogPostSummary";

export interface BlogPostList {
  posts: BlogPostSummary[];
  total: number;
  page: number;
  totalPages: number;
}