# lib/api-zod/src/generated/types/eraseObjectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EraseObjectRequest {
  /** Base64 encoded original image */
  imageBase64: string;
  /** Base64 encoded mask image (white = erase, black = keep) */
  maskBase64: string;
  mimeType: string;
}