# lib/api-zod/src/generated/types/imageBase64Request.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageBase64Request {
  /** Base64 encoded image */
  imageBase64: string;
  /** Image MIME type (image/jpeg, image/png) */
  mimeType: string;
}