# lib/api-zod/src/generated/types/imageBase64Response.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ToolsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageBase64Response {
  /** Base64 encoded result image */
  resultBase64: string;
  mimeType: string;
}